Moved everything from Index to their respective files
