--- src/index.rs
@@ -1,64 +1,0 @@
-use std::ops::Range;
-
-use crossterm::style::Color;
-use rand::{rngs::ThreadRng, thread_rng, Rng};
-
-use crate::*;
-
-/// # "Textbox" struct
-/// Lets you paste a text somewhere in the game screen
-/// 
-/// # DO NOT RELY ON THIS
-/// It'll be replaced in favor of Window system
-/// 
-/// # Warning
-/// The Renderer doesn't check if the text overflows the X position yet, only if it's outside the buffer
-/// 
-/// So be careful where and what you write
-pub struct RENDER_textItem{
-    pub t_position: [usize; 2],
-    pub t_text: String,
-    pub t_lifetime: u16
-}
-
-
-/// # World/Render Buffer Cell
-/// 
-/// Values:
-/// 
-/// * Character
-/// * Color for character
-/// * Color for background
-pub struct TEMPLATE_wrCell {
-    pub c_char: char,
-    pub c_colChr: Color,
-    pub c_colBg: Color,
-}
-impl TEMPLATE_wrCell{
-    pub fn new() -> Self{
-        TEMPLATE_wrCell { c_char: ' ', c_colChr: Color::White, c_colBg: Color::Black }
-    }
-    pub fn newDummy() -> Self{
-        TEMPLATE_wrCell { c_char: '0', c_colChr: Color::Black, c_colBg: Color::White }
-    }
-}
-impl Copy for TEMPLATE_wrCell {}
-impl Clone for TEMPLATE_wrCell {
-    fn clone(&self) -> Self {
-        TEMPLATE_wrCell {
-            c_char: self.c_char,
-            c_colChr: self.c_colChr,
-            c_colBg: self.c_colBg,
-        }
-    }
-}
-
-/// # Interactions enum
-/// # DON'T RELY ON THIS
-/// It will be replaced with introduction of Window system
-pub enum GAME_interactions {
-    i_changeWorldTile,
-    i_printHello,
-    i_printDebug,
-    i_clearWorld,
-}
