+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod renderer;
-pub mod index;
 pub mod system;
 pub mod manufacture;
 pub mod world;
